backend/requirements.txt
absl-py==2.3.1
accelerate==1.9.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
diffusers==0.34.0
distro==1.9.0
filelock==3.18.0
Flask==2.3.3
flask-cors==4.0.0
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
groq==0.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.5
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
jax==0.7.0
jaxlib==0.7.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.14
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.5
numpy==1.24.3
opencv-contrib-python==4.12.0.88
opencv-python==4.8.1.78
opt_einsum==3.4.0
packaging==25.0
Pillow==10.0.1
protobuf==4.25.8
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
safetensors==0.5.3
scipy==1.16.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
